requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
ImageHash==3.7
itsdangerous==0.24
Jinja2==2.9.6
jsonify==0.5
MarkupSafe==1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
scipy==1.0.0
six==1.11.0
uuid==1.30
Werkzeug==0.12.2
